Create another file for data and add a singleton
