+++ pysd/py_backend/components.py
@@ -3,16 +3,10 @@ Model components and time managing classes.
 """
 
 import os
-import warnings
-import re
 import random
 from importlib.machinery import SourceFileLoader
 
-import numpy as np
-import xarray as xr
-
 from pysd._version import __version__
-from .utils import load_outputs, get_columns_to_load
 
 
 class Components(object):
@@ -167,153 +161,3 @@ class Time(object):
     def reset(self):
         """ Reset time value to the initial """
         self._time = self._initial_time
-
-
-class Data(object):
-    # TODO add __init__ and use this clas for used input pandas.Series
-    # as Data
-    # def __init__(self, data, coords, interp="interpolate"):
-
-    def __call__(self, time):
-        try:
-            if time in self.data['time'].values:
-                outdata = self.data.sel(time=time)
-            elif self.interp == "raw":
-                return np.nan
-            elif time > self.data['time'].values[-1]:
-                warnings.warn(
-                    self.py_name + "\n"
-                    + "extrapolating data above the maximum value of the time")
-                outdata = self.data[-1]
-            elif time < self.data['time'].values[0]:
-                warnings.warn(
-                    self.py_name + "\n"
-                    + "extrapolating data below the minimum value of the time")
-                outdata = self.data[0]
-            elif self.interp == "interpolate":
-                outdata = self.data.interp(time=time)
-            elif self.interp == 'look forward':
-                outdata = self.data.sel(time=time, method="backfill")
-            elif self.interp == 'hold backward':
-                outdata = self.data.sel(time=time, method="pad")
-
-            if self.is_float:
-                # if data has no-coords return a float
-                return float(outdata)
-            else:
-                # Remove time coord from the DataArray
-                return outdata.reset_coords('time', drop=True)
-        except Exception as err:
-            if self.data is None:
-                raise ValueError(
-                    self.py_name + "\n"
-                    "Trying to interpolate data variable before loading"
-                    " the data...")
-            else:
-                # raise any other possible error
-                raise err
-
-
-class RegData(Data):
-    def __init__(self, real_name, py_name, coords, interp="interpolate"):
-        self.real_name = real_name
-        self.py_name = py_name
-        self.coords = coords
-        self.interp = interp
-        self.is_float = not bool(coords)
-        self.data = None
-
-    def load_data(self, file_names):
-        """
-        Load data values from files.
-
-        Parameters
-        ----------
-        file_names: list or str
-            Name of the files to search the variable in.
-
-        Returns
-        -------
-        out: xarray.DataArray
-            Resulting data array with the time in the first dimension.
-
-        """
-        if isinstance(file_names, str):
-            file_names = [file_names]
-
-        for file_name in file_names:
-            self.data = self._load_data(file_name)
-            if self.data is not None:
-                break
-
-        if self.data is None:
-            raise ValueError(
-                f"_data_{self.py_name}\n"
-                f"Data for {self.real_name} not found in "
-                f"{', '.join(file_names)}")
-
-    def _load_data(self, file_name):
-        """
-        Load data values from output
-
-        Parameters
-        ----------
-        file_name: str
-            Name of the file to search the variable in.
-
-        Returns
-        -------
-        out: xarray.DataArray or None
-            Resulting data array with the time in the first dimension.
-
-        """
-        # TODO manage missing values on data, as external elements, create a
-        # hierarchy from External?
-        # get columns to load variable
-        columns = get_columns_to_load(
-            file_name, False, vars=[self.real_name, self.py_name])
-
-        if not columns:
-            # try reading transposed file
-            columns = get_columns_to_load(
-                file_name, True, vars=[self.real_name, self.py_name])
-
-            if not columns:
-                # variable not found
-                return None
-            else:
-                # variable must be read from a transposed file
-                transpose = True
-        else:
-            # variable found
-            transpose = False
-
-        if not self.coords:
-            # 0 dimensional data
-            values = load_outputs(file_name, transpose, columns=columns)
-            return xr.DataArray(
-                values.iloc[:, 0].values,
-                {'time': values.index.values},
-                ['time'])
-
-        # subscripted data
-        dims = list(self.coords)
-
-        values = load_outputs(file_name, transpose, columns=columns)
-
-        out = xr.DataArray(
-            np.nan,
-            {'time': values.index.values, **self.coords},
-            ['time'] + dims)
-
-        for column in values.columns:
-            coords = {
-                dim: [coord]
-                for (dim, coord)
-                in zip(dims, re.split(r'\[|\]|\s*,\s*', column)[1:-1])
-            }
-            out.loc[coords] = np.expand_dims(
-                values[column].values,
-                axis=tuple(range(1, len(coords)+1))
-            )
-        return out
