+++ pysd/py_backend/utils.py
@@ -462,82 +462,6 @@ def load_outputs(file_name, transpose=False, columns=None, encoding=None):
         + f"Only {', '.join(list(read_func))} files are accepted.")
 
 
-def get_columns_to_load(file_name, transpose=False, vars=None, encoding=None):
-    """
-    Returns the variable names in the given file.
-
-    Parameters
-    ----------
-    file_name: str
-        Output file to read. Must be csv or tab.
-
-    transpose: bool (optional)
-        If True reads transposed outputs file, i.e. one variable per row.
-        Default is False.
-
-    vars: list or None (optional)
-        List of the vars names to return. If None loads all the column names.
-        Default is None.
-
-    encoding: str or None (optional)
-        Encoding type to read output file. Needed if the file has special
-        characters. Default is None.
-
-    Returns
-    -------
-    set
-        Set of columns for var.
-
-    """
-    read_func = {'.csv': pd.read_csv, '.tab': pd.read_table}
-    out = None
-
-    for end, func in read_func.items():
-        if file_name.lower().endswith(end):
-            if transpose:
-                out = func(file_name,
-                           encoding=encoding,
-                           usecols=[0],
-                           dtype=str).iloc[:, 0].to_list()
-            else:
-                out = func(file_name,
-                           encoding=encoding,
-                           nrows=1,
-                           dtype=str).iloc[:, 1:]
-
-            out = set(out)
-
-    if out is None:
-        raise ValueError(
-            f"\nNot able to read '{file_name}'. "
-            + f"Only {', '.join(list(read_func))} files are accepted.")
-
-    if vars is None:
-        # Not var specified, return all available variables
-        return out
-
-    else:
-        vars_extended = []
-        for var in vars:
-            vars_extended.append(var)
-            if var.startswith('"') and var.endswith('"'):
-                # the variables in "" are reded without " by pandas
-                vars_extended.append(var[1:-1])
-        outs = set()
-        for var in out:
-            if var in vars_extended:
-                # var is in vars_extended (no subscripts)
-                outs.add(var)
-                vars_extended.remove(var)
-            else:
-                for var1 in vars_extended:
-                    if var.startswith(var1 + "["):
-                        # var is subscripted
-                        outs.add(var)
-
-        return outs
-
-
 def detect_encoding(filename):
     """
     Detects the encoding of a file.
